refactor(completion): split shell generators
Break completion generation into per-shell modules plus shared tree/
escaping helpers so future shell work can evolve without a single
monolithic file.

Also route fish/powershell through explicit unsupported stubs and
include related tooling metadata updates in this refactor batch.
