+++ .dprint.jsonc
@@ -33,13 +33,13 @@
 	},
 	"excludes": ["**/node_modules", "**/*-lock.json", "**/*-lock.*", "**/bun.lock*", "**/snapshots"],
 	"plugins": [
-		"https://plugins.dprint.dev/biome-0.11.13.wasm",
+		"https://plugins.dprint.dev/biome-0.12.5.wasm",
 		"https://plugins.dprint.dev/dockerfile-0.3.3.wasm",
-		"https://plugins.dprint.dev/exec-0.6.0.json@a054130d458f124f9b5c91484833828950723a5af3f8ff2bd1523bd47b83b364",
+		"https://plugins.dprint.dev/exec-0.6.2.json@df98f54ffd3092b8a841aedd6d098a2651f16d0a796a40535774f1a8b4b9d463",
 		"https://plugins.dprint.dev/g-plane/malva-v0.15.2.wasm",
-		"https://plugins.dprint.dev/g-plane/markup_fmt-v0.25.3.wasm",
+		"https://plugins.dprint.dev/g-plane/markup_fmt-v0.27.0.wasm",
 		"https://plugins.dprint.dev/g-plane/pretty_yaml-v0.6.0.wasm",
-		"https://plugins.dprint.dev/json-0.21.1.wasm",
+		"https://plugins.dprint.dev/json-0.21.3.wasm",
 		"https://plugins.dprint.dev/markdown-0.21.1.wasm",
 		"https://plugins.dprint.dev/toml-0.7.0.wasm"
 	]
