+++ package.json
@@ -41,6 +41,7 @@
 		"!src/**/*.test.ts"
 	],
 	"scripts": {
+		"bd": "bun run build",
 		"build": "tsdown",
 		"check": "tsgo --noEmit",
 		"check:tsc": "tsc --noEmit",
@@ -73,6 +74,10 @@
 		"bun": ">=1.3",
 		"deno": ">=2.6",
 		"node": ">=22"
+	},
+	"volta": {
+		"node": "25.8.2",
+		"npm": "11.12.1"
 	},
 	"publishConfig": {
 		"exports": {
