+++ src/core/completion/shells/bash.ts
@@ -1,0 +1,320 @@
+/**
+ * Bash completion script generator.
+ *
+ * Produces a self-contained bash script with:
+ * - A main `_<name>_completions` function using `COMP_WORDS`/`COMP_CWORD`
+ * - Multi-level subcommand path detection via progressive COMP_WORDS scanning
+ * - Per-command case branches with flag and subcommand completions
+ * - Propagated flags from ancestor commands at each nesting level
+ * - Enum value completions for `--flag=<value>` patterns
+ * - A `complete -F` registration line
+ *
+ * **Requires** the `bash-completion` package (`_init_completion`).
+ *
+ * @module dreamcli/core/completion/shells/bash
+ * @internal
+ */
+
+import type { CLISchema } from '../../cli/index.ts';
+import type { CommandSchema, FlagSchema } from '../../schema/index.ts';
+import type { CommandNode, CompletionOptions } from './shared.ts';
+import { quoteShellArg, sanitizeShellIdentifier, versionTag, walkCommandTree } from './shared.ts';
+
+// ---------------------------------------------------------------------------
+// Public generator
+// ---------------------------------------------------------------------------
+
+/**
+ * Generates a bash completion script for the CLI.
+ *
+ * @param schema - The CLI schema.
+ * @param options - Optional generator configuration.
+ * @returns A complete bash completion script.
+ */
+function generateBashCompletion(schema: CLISchema, options?: CompletionOptions): string {
+	const prefix = options?.functionPrefix ?? schema.name;
+	const funcName = `_${sanitizeShellIdentifier(prefix)}_completions`;
+	const visibleCommands = schema.commands.map((c) => c.schema).filter((s) => !s.hidden);
+	const nodes = walkCommandTree(visibleCommands);
+	const maxDepth = nodes.reduce((max, n) => Math.max(max, n.path.length), 0);
+
+	const lines: string[] = [];
+
+	lines.push('#!/usr/bin/env bash');
+	lines.push(`# Bash completion for ${schema.name}`);
+	lines.push(`# Generated by ${versionTag()}`);
+	lines.push('#');
+	lines.push(`# Usage:  source <(${schema.name} completions bash)`);
+	lines.push('#');
+	lines.push('# Install permanently:');
+	lines.push(
+		`#   ${schema.name} completions bash > /etc/bash_completion.d/${schema.name}    # Linux`,
+	);
+	lines.push(
+		`#   ${schema.name} completions bash > $(brew --prefix)/etc/bash_completion.d/${schema.name}  # macOS`,
+	);
+	lines.push('#');
+	lines.push('# Requires: bash-completion (provides _init_completion)');
+	lines.push('');
+	lines.push(`${funcName}() {`);
+	lines.push(`\tlocal cur prev words cword`);
+	lines.push(`\t_init_completion || return`);
+	lines.push('');
+
+	// --- Subcommand dispatch ---
+	if (visibleCommands.length > 0) {
+		const groupNodes = nodes.filter((n) => n.children.length > 0);
+
+		// Build subcommand path by scanning COMP_WORDS at each depth level.
+		// At depth 0 we look for top-level command names, at depth 1 we look
+		// for children of the matched depth-0 command, etc.
+		lines.push('\t# Build subcommand path by scanning COMP_WORDS');
+		lines.push('\tlocal subcmd_path=""');
+		lines.push('\tlocal i');
+		lines.push('\tfor ((i = 1; i < cword; i++)); do');
+		// biome-ignore lint/suspicious/noTemplateCurlyInString: bash syntax, not JS template
+		lines.push('\t\tcase "${words[i]}" in');
+		lines.push('\t\t\t-*) continue ;;');
+		lines.push('\t\tesac');
+
+		// Emit progressive path-building cases
+		emitBashPathDetection(lines, visibleCommands, groupNodes, maxDepth);
+
+		lines.push('\tdone');
+		lines.push('');
+
+		// Per-path completions: one case branch per node
+		lines.push('\t# Complete flags and subcommands for the active path');
+		lines.push('\tcase "$subcmd_path" in');
+
+		for (const node of nodes) {
+			const pathKey = node.path.join(' ');
+			const allNames = [node.schema.name, ...node.schema.aliases];
+			const mergedFlags = node.mergedFlags;
+			const enumCases = collectEnumCasesFromFlags(mergedFlags);
+			const flagWords = collectFlagWords(mergedFlags);
+			const childNames = node.children.map((c) => escapeForSingleQuote(c.name)).join(' ');
+			const escapedFlagWords = flagWords
+				.split(' ')
+				.filter(Boolean)
+				.map(escapeForSingleQuote)
+				.join(' ');
+			const completionWords =
+				childNames.length > 0 ? `${childNames} ${escapedFlagWords}` : escapedFlagWords;
+
+			// Use both name and aliases as case patterns via pathKey variations
+			if (node.path.length === 1) {
+				lines.push(`\t\t${allNames.join('|')})`);
+			} else {
+				// For nested paths, the path key always uses the canonical name
+				lines.push(`\t\t${quoteShellCasePattern(pathKey)})`);
+			}
+
+			if (enumCases.length > 0) {
+				lines.push('\t\t\tcase "$prev" in');
+				for (const ec of enumCases) {
+					lines.push(`\t\t\t\t${ec.flags})`);
+					lines.push(formatBashEnumCompletion(ec.values, '\t\t\t\t\t'));
+					lines.push('\t\t\t\t\treturn');
+					lines.push('\t\t\t\t\t;;');
+				}
+				lines.push('\t\t\t\t*)');
+				lines.push('\t\t\t\t\t;;');
+				lines.push('\t\t\tesac');
+			}
+
+			lines.push(`\t\t\tCOMPREPLY=($(compgen -W '${completionWords.trim()}' -- "$cur"))`);
+			lines.push('\t\t\treturn');
+			lines.push('\t\t\t;;');
+		}
+
+		lines.push('\t\t*)');
+		lines.push('\t\t\t;;');
+		lines.push('\tesac');
+		lines.push('');
+
+		// --- Root-level: subcommands + global flags ---
+		const rootFlags = schema.version !== undefined ? '--help --version' : '--help';
+		const subNames = visibleCommands.map((s) => escapeForSingleQuote(s.name)).join(' ');
+		lines.push(`\t# Root-level completions: subcommands and global flags`);
+		lines.push(`\tCOMPREPLY=($(compgen -W '${subNames} ${rootFlags}' -- "$cur"))`);
+	} else {
+		const globalFlags = schema.version !== undefined ? '--help --version' : '--help';
+		lines.push(`\tCOMPREPLY=($(compgen -W '${globalFlags}' -- "$cur"))`);
+	}
+
+	lines.push('}');
+	lines.push('');
+	lines.push(`complete -F ${funcName} ${quoteShellArg(schema.name)}`);
+	lines.push('');
+
+	return lines.join('\n');
+}
+
+// ---------------------------------------------------------------------------
+// Bash helpers — internal
+// ---------------------------------------------------------------------------
+
+/**
+ * Emit bash case branches for progressive subcommand path detection.
+ *
+ * For each nesting depth, generates a `case "$subcmd_path"` that matches
+ * the current accumulated path and extends it when a matching child command
+ * name is found in COMP_WORDS.
+ *
+ * @internal
+ */
+function emitBashPathDetection(
+	lines: string[],
+	visibleCommands: readonly CommandSchema[],
+	groupNodes: readonly CommandNode[],
+	maxDepth: number,
+): void {
+	// Depth 0: match top-level command names
+	const topLevelPatterns = visibleCommands.flatMap((s) => [s.name, ...s.aliases]);
+	lines.push(`\t\tcase "$subcmd_path" in`);
+	lines.push(`\t\t\t'')`);
+	lines.push(`\t\t\t\tcase "\${words[i]}" in`);
+	lines.push(`\t\t\t\t\t${topLevelPatterns.join('|')})`);
+	lines.push(`\t\t\t\t\t\tsubcmd_path="\${words[i]}"`);
+	lines.push(`\t\t\t\t\t\t;;`);
+	lines.push(`\t\t\t\tesac`);
+	lines.push(`\t\t\t\t;;`);
+
+	// Deeper levels: for each group that has children, match its path and extend
+	if (maxDepth > 1) {
+		for (const node of groupNodes) {
+			const pathKey = node.path.join(' ');
+			const childPatterns = node.children.flatMap((c) => [c.name, ...c.aliases]);
+
+			lines.push(`\t\t\t${quoteShellCasePattern(pathKey)})`);
+			lines.push(`\t\t\t\tcase "\${words[i]}" in`);
+			lines.push(`\t\t\t\t\t${childPatterns.join('|')})`);
+			lines.push(`\t\t\t\t\t\tsubcmd_path="$subcmd_path \${words[i]}"`);
+			lines.push(`\t\t\t\t\t\t;;`);
+			lines.push(`\t\t\t\tesac`);
+			lines.push(`\t\t\t\t;;`);
+		}
+	}
+
+	lines.push(`\t\tesac`);
+}
+
+/**
+ * Quote a string for use as a bash `case` pattern. Patterns containing
+ * spaces need quoting to match as a single string.
+ *
+ * @internal
+ */
+function quoteShellCasePattern(pattern: string): string {
+	if (!pattern.includes(' ')) return pattern;
+	return `"${pattern}"`;
+}
+
+/**
+ * Escape a string for safe embedding inside an existing single-quoted
+ * shell literal (e.g. inside `compgen -W '...'`).
+ *
+ * Uses the `'\''` idiom: close quote, escaped literal quote, reopen
+ * quote. Returns the input unchanged when it consists only of shell-safe
+ * characters, which covers the vast majority of flag/command names.
+ *
+ * @internal
+ */
+function escapeForSingleQuote(value: string): string {
+	if (/^[a-zA-Z0-9_\-.]+$/.test(value)) return value;
+	return value.replace(/'/g, "'\\''");
+}
+
+/**
+ * Format a COMPREPLY line for bash enum value completion.
+ *
+ * When all values are simple (no whitespace or special chars), emits a
+ * standard `compgen -W 'v1 v2' -- "$cur"` line.
+ *
+ * When any value contains whitespace, single quotes, or backslashes,
+ * uses `IFS=$'\n'` with `$'...'` quoting to split on newlines
+ * instead of spaces, handling all special characters correctly.
+ *
+ * @internal
+ */
+function formatBashEnumCompletion(values: readonly string[], indent: string): string {
+	const needsDollarQuoting = values.some((v) => !/^[a-zA-Z0-9_\-.]+$/.test(v));
+	if (needsDollarQuoting) {
+		const escaped = values.map(escapeBashDollarQuote);
+		const joined = escaped.join('\\n');
+		return `${indent}local IFS=$'\\n'\n${indent}COMPREPLY=($(compgen -W $'${joined}' -- "$cur"))`;
+	}
+	return `${indent}COMPREPLY=($(compgen -W '${values.join(' ')}' -- "$cur"))`;
+}
+
+/**
+ * Escape a single enum value for use inside a bash `$'...'` string.
+ *
+ * In `$'...'` context:
+ * - `\\` → literal backslash
+ * - `\'` → literal single quote
+ * - `\n` → newline (used as IFS delimiter, so must be escaped in values)
+ *
+ * Values that are simple identifiers pass through unescaped.
+ *
+ * @internal
+ */
+function escapeBashDollarQuote(value: string): string {
+	return value.replace(/\\/g, '\\\\').replace(/'/g, "\\'").replace(/\n/g, '\\n');
+}
+
+/**
+ * Collect all `--flag` words (long + short aliases) for a flag record.
+ *
+ * @internal
+ */
+function collectFlagWords(flags: Readonly<Record<string, FlagSchema>>): string {
+	const words: string[] = [];
+	for (const [name, schema] of Object.entries(flags)) {
+		words.push(`--${name}`);
+		for (const alias of schema.aliases) {
+			words.push(alias.length === 1 ? `-${alias}` : `--${alias}`);
+		}
+	}
+	return words.join(' ');
+}
+
+/**
+ * A consolidated enum-value case: the flag forms that trigger it and the values.
+ *
+ * @internal
+ */
+interface EnumCase {
+	readonly flags: string;
+	readonly values: readonly string[];
+}
+
+/**
+ * Collect enum flag cases from a merged flag record.
+ *
+ * Each enum flag produces one case entry with its long/short flag forms
+ * as the pattern and enum values as the completions.
+ *
+ * @internal
+ */
+function collectEnumCasesFromFlags(
+	flags: Readonly<Record<string, FlagSchema>>,
+): readonly EnumCase[] {
+	const cases: EnumCase[] = [];
+	for (const [name, schema] of Object.entries(flags)) {
+		if (schema.kind !== 'enum' || schema.enumValues === undefined) continue;
+		const flagForms = [
+			`--${name}`,
+			...schema.aliases.map((a) => (a.length === 1 ? `-${a}` : `--${a}`)),
+		];
+		cases.push({ flags: flagForms.join('|'), values: schema.enumValues });
+	}
+	return cases;
+}
+
+// ---------------------------------------------------------------------------
+// Exports
+// ---------------------------------------------------------------------------
+
+export { generateBashCompletion };
