+++ src/core/completion/shells/zsh.ts
@@ -1,0 +1,325 @@
+/**
+ * Zsh completion script generator.
+ *
+ * Produces a self-contained zsh script with:
+ * - A `#compdef` directive for auto-loading
+ * - A main `_<name>` function with subcommand dispatch
+ * - Per-group helper functions (`_<name>_<cmd>`) with recursive dispatch
+ * - `_arguments` calls for flags with descriptions (including propagated flags)
+ * - `_describe` for subcommand lists at each nesting level
+ * - Enum value completions via `->state` dispatch
+ *
+ * @module dreamcli/core/completion/shells/zsh
+ * @internal
+ */
+
+import type { CLISchema } from '../../cli/index.ts';
+import type { CommandSchema, FlagSchema } from '../../schema/index.ts';
+import type { CommandNode, CompletionOptions } from './shared.ts';
+import { quoteShellArg, sanitizeShellIdentifier, versionTag, walkCommandTree } from './shared.ts';
+
+// ---------------------------------------------------------------------------
+// Public generator
+// ---------------------------------------------------------------------------
+
+/**
+ * Generates a zsh completion script for the CLI.
+ *
+ * @param schema - The CLI schema.
+ * @param options - Optional generator configuration.
+ * @returns A complete zsh completion script.
+ */
+function generateZshCompletion(schema: CLISchema, options?: CompletionOptions): string {
+	const prefix = options?.functionPrefix ?? schema.name;
+	const funcName = `_${sanitizeShellIdentifier(prefix)}`;
+	const visibleCommands = schema.commands.map((c) => c.schema).filter((s) => !s.hidden);
+	const nodes = walkCommandTree(visibleCommands);
+
+	const lines: string[] = [];
+
+	lines.push(`#compdef ${quoteShellArg(schema.name)}`);
+	lines.push(`# Zsh completion for ${schema.name}`);
+	lines.push(`# Generated by ${versionTag()}`);
+	lines.push('#');
+	lines.push(`# Usage:  source <(${schema.name} completions zsh)`);
+	lines.push('#');
+	lines.push('# Install permanently:');
+	lines.push(`#   ${schema.name} completions zsh > "\${fpath[1]}/_${schema.name}"`);
+	lines.push('#   # Then restart your shell');
+	lines.push('');
+
+	// --- Generate helper functions for group commands (bottom-up) ---
+	// Process deepest nodes first so all helpers exist before they're called.
+	const groupNodes = nodes.filter((n) => n.children.length > 0);
+	const leafNodes = nodes.filter((n) => n.children.length === 0);
+
+	// Emit leaf helper functions
+	for (const node of leafNodes) {
+		if (node.path.length < 2) continue; // Top-level leaves handled inline
+		const helperName = `${funcName}_${node.path.map(sanitizeShellIdentifier).join('_')}`;
+		emitZshLeafFunction(lines, helperName, node);
+	}
+
+	// Emit group helper functions (deepest first)
+	const sortedGroups = [...groupNodes].sort((a, b) => b.path.length - a.path.length);
+	for (const node of sortedGroups) {
+		const helperName = `${funcName}_${node.path.map(sanitizeShellIdentifier).join('_')}`;
+		emitZshGroupFunction(lines, helperName, funcName, node);
+	}
+
+	// --- Main function ---
+	lines.push(`${funcName}() {`);
+
+	if (visibleCommands.length > 0) {
+		lines.push('\tlocal -a subcmds');
+		lines.push('\tlocal line state');
+		lines.push('');
+
+		// --- Root-level: global flags + subcommand argument ---
+		lines.push('\t_arguments -C \\');
+		lines.push("\t\t'--help[Show help text]' \\");
+		if (schema.version !== undefined) {
+			lines.push("\t\t'--version[Show version]' \\");
+		}
+		lines.push("\t\t'1: :->subcmd' \\");
+		lines.push("\t\t'*::arg:->args'");
+		lines.push('');
+
+		// --- Subcommand list ---
+		lines.push('\tcase "$state" in');
+		lines.push('\t\tsubcmd)');
+		lines.push('\t\t\tsubcmds=(');
+		for (const cmd of visibleCommands) {
+			const desc = escapeZshDescription(cmd.description ?? cmd.name);
+			lines.push(`\t\t\t\t'${cmd.name}:${desc}'`);
+		}
+		lines.push('\t\t\t)');
+		lines.push("\t\t\t_describe 'command' subcmds");
+		lines.push('\t\t\t;;');
+
+		// --- Per-command arguments ---
+		lines.push('\t\targs)');
+		lines.push('\t\t\tcase "$line[1]" in');
+		for (const cmd of visibleCommands) {
+			const allNames = [cmd.name, ...cmd.aliases];
+			const childCommands = cmd.commands.filter((c) => !c.hidden);
+			lines.push(`\t\t\t\t${allNames.join('|')})`);
+
+			if (childCommands.length > 0) {
+				// Group command — delegate to helper function
+				const helperName = `${funcName}_${sanitizeShellIdentifier(cmd.name)}`;
+				lines.push(`\t\t\t\t\t${helperName}`);
+			} else {
+				// Leaf command — emit flags inline
+				const node = nodes.find((n) => n.path.length === 1 && n.schema.name === cmd.name);
+				const flagSpecs = node
+					? buildZshFlagSpecsFromFlags(node.mergedFlags)
+					: buildZshFlagSpecs(cmd);
+
+				if (flagSpecs.length > 0) {
+					lines.push('\t\t\t\t\t_arguments \\');
+					for (let i = 0; i < flagSpecs.length; i++) {
+						const trailing = i < flagSpecs.length - 1 ? ' \\' : '';
+						lines.push(`\t\t\t\t\t\t${flagSpecs[i]}${trailing}`);
+					}
+				}
+			}
+			lines.push('\t\t\t\t\t;;');
+		}
+		lines.push('\t\t\tesac');
+		lines.push('\t\t\t;;');
+		lines.push('\tesac');
+	} else {
+		// --- No subcommands: just global flags ---
+		const globalFlags: string[] = ["'--help[Show help text]'"];
+		if (schema.version !== undefined) {
+			globalFlags.push("'--version[Show version]'");
+		}
+		lines.push('\t_arguments \\');
+		for (let i = 0; i < globalFlags.length; i++) {
+			const trailing = i < globalFlags.length - 1 ? ' \\' : '';
+			lines.push(`\t\t${globalFlags[i]}${trailing}`);
+		}
+	}
+
+	lines.push('}');
+	lines.push('');
+	lines.push(`${funcName} "$@"`);
+	lines.push('');
+
+	return lines.join('\n');
+}
+
+// ---------------------------------------------------------------------------
+// Zsh helpers — internal
+// ---------------------------------------------------------------------------
+
+/**
+ * Emit a zsh helper function for a group command (has subcommands).
+ *
+ * Uses `_arguments -C` for its own flags plus subcommand dispatch.
+ *
+ * @internal
+ */
+function emitZshGroupFunction(
+	lines: string[],
+	helperName: string,
+	rootFuncName: string,
+	node: CommandNode,
+): void {
+	lines.push(`${helperName}() {`);
+	lines.push('\tlocal -a subcmds');
+	lines.push('\tlocal line state');
+	lines.push('');
+
+	// Flags for this group (including propagated)
+	const flagSpecs = buildZshFlagSpecsFromFlags(node.mergedFlags);
+
+	lines.push('\t_arguments -C \\');
+	for (const spec of flagSpecs) {
+		lines.push(`\t\t${spec} \\`);
+	}
+	lines.push("\t\t'1: :->subcmd' \\");
+	lines.push("\t\t'*::arg:->args'");
+	lines.push('');
+
+	lines.push('\tcase "$state" in');
+	lines.push('\t\tsubcmd)');
+	lines.push('\t\t\tsubcmds=(');
+	for (const child of node.children) {
+		const desc = escapeZshDescription(child.description ?? child.name);
+		lines.push(`\t\t\t\t'${child.name}:${desc}'`);
+	}
+	lines.push('\t\t\t)');
+	lines.push("\t\t\t_describe 'command' subcmds");
+	lines.push('\t\t\t;;');
+
+	lines.push('\t\targs)');
+	lines.push('\t\t\tcase "$line[1]" in');
+	for (const child of node.children) {
+		const allNames = [child.name, ...child.aliases];
+		const childPath = [...node.path, child.name];
+		const childFuncName = `${rootFuncName}_${childPath.map(sanitizeShellIdentifier).join('_')}`;
+		lines.push(`\t\t\t\t${allNames.join('|')})`);
+		lines.push(`\t\t\t\t\t${childFuncName}`);
+		lines.push('\t\t\t\t\t;;');
+	}
+	lines.push('\t\t\tesac');
+	lines.push('\t\t\t;;');
+	lines.push('\tesac');
+
+	lines.push('}');
+	lines.push('');
+}
+
+/**
+ * Emit a zsh helper function for a leaf command (no subcommands).
+ *
+ * Uses plain `_arguments` with the command's own flags plus propagated flags.
+ *
+ * @internal
+ */
+function emitZshLeafFunction(lines: string[], helperName: string, node: CommandNode): void {
+	const flagSpecs = buildZshFlagSpecsFromFlags(node.mergedFlags);
+
+	lines.push(`${helperName}() {`);
+	if (flagSpecs.length > 0) {
+		lines.push('\t_arguments \\');
+		for (let i = 0; i < flagSpecs.length; i++) {
+			const trailing = i < flagSpecs.length - 1 ? ' \\' : '';
+			lines.push(`\t\t${flagSpecs[i]}${trailing}`);
+		}
+	}
+	lines.push('}');
+	lines.push('');
+}
+
+/**
+ * Escape single quotes and colons in descriptions for zsh `_describe` format.
+ * Zsh uses `'name:description'` syntax — colons in descriptions must be escaped
+ * with backslash, and single quotes must be handled.
+ *
+ * @internal
+ */
+function escapeZshDescription(text: string): string {
+	return text.replace(/\\/g, '\\\\').replace(/'/g, "'\\''").replace(/:/g, '\\:');
+}
+
+/**
+ * Escape a single enum value for use inside a zsh `(v1 v2)` value list.
+ *
+ * Simple identifier-like values (matching `[a-zA-Z0-9_\-.]`) pass through
+ * unescaped. Values containing spaces, quotes, backslashes, or parentheses
+ * are escaped with backslash prefixes for each special character, which
+ * is the standard mechanism within zsh parenthesized completion lists.
+ *
+ * @internal
+ */
+function escapeZshEnumValue(value: string): string {
+	if (/^[a-zA-Z0-9_\-.]+$/.test(value)) return value;
+	return value
+		.replace(/\\/g, '\\\\')
+		.replace(/'/g, "\\'")
+		.replace(/ /g, '\\ ')
+		.replace(/\(/g, '\\(')
+		.replace(/\)/g, '\\)');
+}
+
+/**
+ * Build `_arguments` flag spec strings from a merged flag record.
+ *
+ * Each flag produces one or more spec strings in zsh `_arguments` format:
+ * - `'--name[description]'` for simple flags
+ * - `'--name[description]:value:(v1 v2 v3)'` for enum flags
+ * - `'(-s --name)'{-s,--name}'[description]'` for flags with short aliases
+ *
+ * @internal
+ */
+function buildZshFlagSpecsFromFlags(
+	flags: Readonly<Record<string, FlagSchema>>,
+): readonly string[] {
+	const specs: string[] = [];
+	for (const [name, schema] of Object.entries(flags)) {
+		const desc = escapeZshDescription(schema.description ?? name);
+		const longFlag = `--${name}`;
+		const shortFlags = schema.aliases.filter((a) => a.length === 1).map((a) => `-${a}`);
+
+		let valuePart = '';
+		if (schema.kind === 'boolean') {
+			// Boolean flags take no argument value
+		} else if (schema.kind === 'enum' && schema.enumValues !== undefined) {
+			const escapedValues = schema.enumValues.map(escapeZshEnumValue);
+			valuePart = `:value:(${escapedValues.join(' ')})`;
+		} else {
+			valuePart = ':value:';
+		}
+
+		if (shortFlags.length > 0) {
+			// Mutual exclusion group with ALL short aliases + long flag
+			// Format: '(-v -V --verbose)'{-v,-V,--verbose}'[desc]:value'
+			const allForms = [...shortFlags, longFlag];
+			const excl = `(${allForms.join(' ')})`;
+			specs.push(`'${excl}'{${allForms.join(',')}}'[${desc}]${valuePart}'`);
+		} else {
+			specs.push(`'${longFlag}[${desc}]${valuePart}'`);
+		}
+	}
+	return specs;
+}
+
+/**
+ * Build `_arguments` flag spec strings for a command.
+ *
+ * Delegates to {@link buildZshFlagSpecsFromFlags} using the command's own flags.
+ *
+ * @internal
+ */
+function buildZshFlagSpecs(cmd: CommandSchema): readonly string[] {
+	return buildZshFlagSpecsFromFlags(cmd.flags);
+}
+
+// ---------------------------------------------------------------------------
+// Exports
+// ---------------------------------------------------------------------------
+
+export { generateZshCompletion };
